JavaScreen caching code is largely implemented except for checking the validity of the cache files; committing with caching disabled to work on cursor draw command ordering (includes improvements to DevFileHeader class).

+++ graphics.new/JavaScreenCmd.C
@@ -21,6 +21,10 @@
   $Id$
 
   $Log$
+  Revision 1.99  2000/03/28 21:46:25  wenger
+  Started implementation of caching commands, GIFs, and GData to speed up
+  session opening.
+
   Revision 1.98  2000/03/27 21:42:25  wenger
   Removed unused JavaScreenCmd::DrawAllCursors() method.
 
@@ -447,6 +451,7 @@
 #include "PileStack.h"
 #include "DebugLog.h"
 #include "ElapsedTime.h"
+#include "DevFileHeader.h"
 
 //#define DEBUG
 #define DEBUG_LOG
@@ -479,6 +484,10 @@ static const float viewZInc = 0.001;
 
 static const int protocolMajorVersion = 4;
 static const int protocolMinorVersion = 1;
+
+static const char *_dispSizePrefix = "display size: ";
+static const char *_protocolVersionPrefix = "protocol version: ";
+static const char *_cmdPrefix = "command: ";
 
 // be very careful that this order agree with the ControlCmdType definition
 char* JavaScreenCmd::_controlCmdName[JavaScreenCmd::CONTROLCMD_NUM]=
@@ -938,7 +947,11 @@ JavaScreenCmd::JavaScreenCmd(ControlPanel* control,
 
 	_recording = false;
 	_commandFile = NULL;
+	_commandFileName = NULL;
 	_dataFile = -1;
+	_dataFileName = NULL;
+
+	_playingBack = false;
 
     for (i=0; i< _argc; ++i)
     {
@@ -1151,6 +1164,10 @@ JavaScreenCmd::DoOpenSession(char *fullpath)
 	// and other problems.
 	DoCloseSession();
 
+#if 0//TEMPTEMP // Under development.
+    (void)StartPlayingBack(fullpath);
+#endif
+
 	_postponeCursorCmds = true;
 
 	// Set modes, etc., for session to work correctly in JavaScreen.
@@ -1225,8 +1242,10 @@ JavaScreenCmd::DoOpenSession(char *fullpath)
 	// ...end of kludgey section.
 	//
 
-#if 0 // Under development.
-	(void)StartRecording(fullpath);
+#if 0//TEMPTEMP // Under development.
+	if (!_playingBack) {
+	    (void)StartRecording(fullpath);
+	}
 #endif
 
 	// Send commands to create the top-level views.
@@ -1258,9 +1277,13 @@ JavaScreenCmd::DoOpenSession(char *fullpath)
 		if (_recording) _recordingStatus += StatusFailed;
     }
 
-#if 0 // Under development.
-	(void)StopRecording();
-#endif
+	if (_recording) {
+	    (void)StopRecording();
+	}
+
+	if (_playingBack) {
+	    (void)StopPlayingBack();
+	}
 
 #if JS_TIMER
 	et.ReportTime("OpenSession");
@@ -1902,6 +1925,10 @@ JavaScreenCmd::SendWindowData(const char* fileName, Boolean doChecksum,
 	return status;
 #endif
 
+	// If we're in playback mode, the data will have been sent by the
+	// playback code, so don't do anything here.
+    if (_playingBack) return status;
+
 	int filesize = 0;
 
 	// send the dumped image file to the JAVA_SCREEN client
@@ -2301,7 +2328,9 @@ JavaScreenCmd::RequestUpdateGData(ViewGraph *view)
 
 //====================================================================
 void
-JavaScreenCmd::DrawCursor(View *view, DeviseCursor *cursor)
+//TEMPTEMP -- jsc argument here is a kludge -- _playingBack, etc., should
+// probably be static
+JavaScreenCmd::DrawCursor(View *view, DeviseCursor *cursor, JavaScreenCmd *jsc)
 {
 #if defined (DEBUG_LOG)
     sprintf(logBuf, "JavaScreenCmd::DrawCursor(%s, %s)\n", view->GetName(),
@@ -2426,8 +2455,13 @@ JavaScreenCmd::DrawCursor(View *view, DeviseCursor *cursor)
 	//
     // This JavaScreenCmd object is created only to send the command.
 	//
+#if 0 //TEMPTEMP
+	//TEMPTEMP -- this object doesn't have recording turned on
     JavaScreenCmd jsc(ControlPanel::Instance(), NULL_SVC_CMD, 0, NULL);
 	args.ReturnVal(&jsc);
+#else //TEMPTEMP
+	args.ReturnVal(jsc);
+#endif //TEMPTEMP
 }
 
 //====================================================================
@@ -2481,10 +2515,17 @@ JavaScreenCmd::ReturnVal(int argc, char** argv)
     printf(")\n");
 #endif
 
+	// If we're in playback mode, the command or result will have been sent
+	// by the playback code, so don't do anything here.
+    if (_playingBack) return 0;
+
     // record the command in the command file if necessary
 	if (_recording) {
-	    fprintf(_commandFile, "command: ");
-		PrintArgs(_commandFile, argc, argv, true);//TEMP?
+	    fprintf(_commandFile, _cmdPrefix);
+		for (int aNum = 0; aNum < argc; aNum++) {
+		    fprintf(_commandFile, "{%s} ", argv[aNum]);
+		}
+		fprintf(_commandFile, "\n");
 	}
 
 	// send the command out
@@ -2643,7 +2684,7 @@ JavaScreenCmd::DrawChangedCursors()
 		DeviseCursor *cursor = _drawnCursors.Next(index);
 		View *view = cursor->GetDst();
 		if (view != NULL) {
-			DrawCursor(view, cursor);
+			DrawCursor(view, cursor, this);
 		}
 	}
 	_drawnCursors.DoneIterator(index);
@@ -3009,6 +3050,128 @@ JavaScreenCmd::UpdateViewImage(View *view, int imageSize)
 
 //====================================================================
 DevStatus
+JavaScreenCmd::StartPlayingBack(const char *sessionFile)
+{
+#if defined (DEBUG_LOG)
+    sprintf(logBuf, "JavaScreenCmd::StartPlayingBack(%s)\n", sessionFile);
+    DebugLog::DefaultLog()->Message(DebugLog::LevelInfo1, logBuf);
+#endif
+
+    DevStatus result(StatusOk);
+
+//TEMPTEMP -- make sure cache files are newer than session file
+
+    result += OpenCacheFiles(sessionFile, false);
+
+	if (result.IsComplete()) {
+
+		// Check file header.
+		DevFileHeader::Info info;
+		//TEMPTEMP -- check result
+	    result += DevFileHeader::Read(_commandFile, info);
+
+	    if (result.IsComplete()) {
+		    if (strcmp(info.fileType, FILE_TYPE_JSCMDCACHE)) {
+				char errBuf[MAXPATHLEN + 128];
+				sprintf(errBuf,
+				  "File %s is not a JavaScreen command cache file",
+				  _commandFileName);
+				reportErrNosys(errBuf);
+			    result += StatusFailed;
+			}
+		}
+	}
+
+	if (result.IsComplete()) {
+//TEMPTEMP -- check protocol version
+
+        // Check display size for match
+		int prefixLen = strlen(_dispSizePrefix);
+		const int sizeBufSize = 512;
+        char sizeBuf[sizeBufSize];
+		Boolean foundSize = false;
+		//TEMPTEMP -- use nice_fgets??
+		//TEMPTEMP - what if we don't find it at all?
+		while (fgets(sizeBuf, sizeBufSize, _commandFile) && !foundSize) {
+		    if (!strncmp(sizeBuf, _dispSizePrefix, prefixLen)) {
+			    foundSize = true;
+				int cacheWidth, cacheHeight;
+				sscanf(&sizeBuf[prefixLen], "%dx%d", &cacheWidth,
+				  &cacheHeight);
+				int curWidth =
+				  DeviseDisplay::DefaultDisplay()->DesiredScreenWidth();
+				int curHeight =
+				  DeviseDisplay::DefaultDisplay()->DesiredScreenHeight();
+				if (cacheWidth != curWidth || cacheHeight != curHeight) {
+				    char errBuf[1024];
+					sprintf(errBuf, "Current display size %dx%d does not match cache file display size of %dx%d; cached data not used",
+					  curWidth, curHeight, cacheWidth, cacheHeight);
+					reportErrNosys(errBuf);
+					result += StatusFailed;
+				}
+			}
+	    }
+    }
+
+	if (result.IsComplete()) {
+		// Play back the command file.
+		int prefixLen = strlen(_cmdPrefix);
+		const int cmdBufSize = 512;
+        char cmdBuf[cmdBufSize];
+		//TEMPTEMP -- use nice_fgets??
+		while (fgets(cmdBuf, cmdBufSize, _commandFile)) {
+		    if (!strncmp(cmdBuf, _cmdPrefix, prefixLen)) {
+			    result += SendCmd(&cmdBuf[prefixLen]);
+			}
+		}
+    }
+
+	if (result.IsComplete()) {
+        // Play back the data (image and GData) file.
+		int tmp;
+		//TEMPTEMP -- check return value
+		//TEMPTEMP -- what about checksum?
+		//TEMPTEMP -- maybe don't use SendWindowData here
+	    SendWindowData(_dataFileName, false, tmp);
+
+		// This must be set *after* sending the data, otherwise
+		// SendWindowData() won't do anything.
+	    _playingBack = true;
+    }
+
+//TEMPTEMP -- if we send JAVAC_Done here, what happens if user does
+// something while devised is still really opening the session?
+
+    if (!_playingBack) {
+	    result += CloseCacheFiles(false);
+	}
+
+    return result;
+}
+
+//====================================================================
+DevStatus
+JavaScreenCmd::SendCmd(const char* cmd)
+{
+    DevStatus result(StatusOk);
+
+    ArgList args;
+	args.ParseString(cmd);
+
+#if defined(DEBUG)
+	printf("Sending command: ");
+	PrintArgs(stdout, args.GetCount(), args.GetArgs(), true);
+#endif
+
+	//TEMPTEMP -- check return value
+	//TEMPTEMP -- eliminate cast
+	ReturnVal(args.GetCount(), (char **)args.GetArgs());
+
+    return result;
+}
+
+//====================================================================
+DevStatus
 JavaScreenCmd::StartRecording(const char *sessionFile)
 {
 #if defined (DEBUG_LOG)
@@ -3030,7 +3193,19 @@ JavaScreenCmd::StartRecording(const char *sessionFile)
 	if (result.IsComplete()) {
 	    _recording = true;
 		_recordingStatus = StatusOk;
-		//TEMP -- save display size
+
+		// Write file header.
+		const char *header = DevFileHeader::Get(FILE_TYPE_JSCMDCACHE);
+		fprintf(_commandFile, "%s\n", header);
+
+		// Save display size.
+	    int width = DeviseDisplay::DefaultDisplay()->DesiredScreenWidth();
+	    int height = DeviseDisplay::DefaultDisplay()->DesiredScreenHeight();
+		fprintf(_commandFile, "%s%dx%d\n", _dispSizePrefix, width, height);
+
+		// Save protocol version.
+		fprintf(_commandFile, "%s%d.%d\n\n", _protocolVersionPrefix,
+		  protocolMajorVersion, protocolMinorVersion);
 	} else {
 	    if (_commandFile) {
 		    (void)fclose(_commandFile);
@@ -3039,7 +3214,6 @@ JavaScreenCmd::StartRecording(const char *sessionFile)
 	}
 
 //TEMP -- what if command file already exists?
-//TEMP -- add file header?
 
     return result;
 }
@@ -3049,6 +3223,9 @@ DevStatus
 JavaScreenCmd::OpenCacheFiles(const char *sessionFile, Boolean writing)
 {
     DevStatus result(StatusOk);
+
+	_commandFile = NULL;
+	_dataFile = -1;
 
 	if (result.IsComplete()) {
 		char fileBuf[MAXPATHLEN];
@@ -3075,7 +3252,35 @@ JavaScreenCmd::OpenCacheFiles(const char *sessionFile, Boolean writing)
 		}
 	}
 
+	if (!result.IsComplete()) {
+	    result += CloseCacheFiles(false);
+	}
+
 	return result;
+}
+
+//====================================================================
+DevStatus
+JavaScreenCmd::StopPlayingBack()
+{
+#if defined (DEBUG_LOG)
+    sprintf(logBuf, "JavaScreenCmd::StopPlayingBack()\n");
+    DebugLog::DefaultLog()->Message(DebugLog::LevelInfo1, logBuf);
+#endif
+
+    DevStatus result(StatusOk);
+
+    if (!_playingBack) {
+	    //TEMPTEMP
+	}
+
+	if (result.IsComplete()) {
+		result += CloseCacheFiles(false);
+	}
+
+	_playingBack = false;
+
+    return result;
 }
 
 //====================================================================
@@ -3107,13 +3312,15 @@ JavaScreenCmd::CloseCacheFiles(Boolean deleteFiles)
 {
     DevStatus result(StatusOk);
 
-	if (fclose(_commandFile) != 0) {
-	    reportErrSys("Error closing command file");
-	    result = StatusFailed;
+	if (_commandFile) {
+	    if (fclose(_commandFile) != 0) {
+	        reportErrSys("Error closing command file");
+	        result = StatusFailed;
+	    }
+	    _commandFile = NULL;
 	}
-	_commandFile = NULL;
 
-	if (deleteFiles) {
+	if (deleteFiles && _commandFileName) {
 	    if (unlink(_commandFileName) < 0) {
 		    sprintf(logBuf, "Error unlinking command file %s",
 			  _commandFileName);
@@ -3124,13 +3331,15 @@ JavaScreenCmd::CloseCacheFiles(Boolean deleteFiles)
 	FreeString(_commandFileName);
 	_commandFileName = NULL;
 
-	if (close(_dataFile) < 0) {
-	    reportErrSys("Error closing data file");
-		result = StatusFailed;
+	if (_dataFile != -1) {
+	    if (close(_dataFile) < 0) {
+	        reportErrSys("Error closing data file");
+		    result = StatusFailed;
+	    }
+	    _dataFile = -1;
 	}
-	_dataFile = -1;
 
-	if (deleteFiles) {
+	if (deleteFiles && _dataFileName) {
 	    if (unlink(_dataFileName) < 0) {
 		    sprintf(logBuf, "Error unlinking data file %s", _dataFileName);
 		    reportErrSys(logBuf);
